+++ lib/zendesk2/help_center.rb
@@ -1,27 +1,29 @@
 module Zendesk2::HelpCenter; end
 
-require 'zendesk2/requests/create_help_center_article.rb'
-require 'zendesk2/requests/create_help_center_category.rb'
-require 'zendesk2/requests/create_help_center_section.rb'
-require 'zendesk2/requests/destroy_help_center_article.rb'
-require 'zendesk2/requests/destroy_help_center_category.rb'
-require 'zendesk2/requests/destroy_help_center_section.rb'
-require 'zendesk2/requests/get_help_center_article.rb'
-require 'zendesk2/requests/get_help_center_articles.rb'
-require 'zendesk2/requests/get_help_center_categories.rb'
-require 'zendesk2/requests/get_help_center_categories_articles.rb'
-require 'zendesk2/requests/get_help_center_categories_sections.rb'
-require 'zendesk2/requests/get_help_center_category.rb'
-require 'zendesk2/requests/get_help_center_section.rb'
-require 'zendesk2/requests/get_help_center_sections.rb'
-require 'zendesk2/requests/get_help_center_sections_articles.rb'
-require 'zendesk2/requests/search_help_center_articles.rb'
-require 'zendesk2/requests/update_help_center_article.rb'
-require 'zendesk2/requests/update_help_center_category.rb'
-require 'zendesk2/requests/update_help_center_section.rb'
-require 'zendesk2/models/help_center/article.rb'
-require 'zendesk2/models/help_center/category.rb'
-require 'zendesk2/models/help_center/section.rb'
-require 'zendesk2/collections/help_center/articles.rb'
-require 'zendesk2/collections/help_center/categories.rb'
-require 'zendesk2/collections/help_center/sections.rb'
+require 'zendesk2/help_center/create_help_center_article'
+require 'zendesk2/help_center/create_help_center_category'
+require 'zendesk2/help_center/create_help_center_section'
+require 'zendesk2/help_center/destroy_help_center_article'
+require 'zendesk2/help_center/destroy_help_center_category'
+require 'zendesk2/help_center/destroy_help_center_section'
+require 'zendesk2/help_center/get_help_center_article'
+require 'zendesk2/help_center/get_help_center_articles'
+require 'zendesk2/help_center/get_help_center_categories'
+require 'zendesk2/help_center/get_help_center_categories_articles'
+require 'zendesk2/help_center/get_help_center_categories_sections'
+require 'zendesk2/help_center/get_help_center_category'
+require 'zendesk2/help_center/get_help_center_section'
+require 'zendesk2/help_center/get_help_center_sections'
+require 'zendesk2/help_center/get_help_center_sections_articles'
+require 'zendesk2/help_center/search_help_center_articles'
+require 'zendesk2/help_center/update_help_center_article'
+require 'zendesk2/help_center/update_help_center_category'
+require 'zendesk2/help_center/update_help_center_section'
+
+require 'zendesk2/help_center/article'
+require 'zendesk2/help_center/category'
+require 'zendesk2/help_center/section'
+
+require 'zendesk2/help_center/articles'
+require 'zendesk2/help_center/categories'
+require 'zendesk2/help_center/sections'
