+++ lib/zendesk2/help_center/article.rb
