+++ lib/zendesk2/help_center/categories.rb
