+++ lib/zendesk2/help_center/category.rb
