+++ lib/zendesk2/help_center/create_help_center_section.rb
