+++ lib/zendesk2/help_center/destroy_help_center_section.rb
