+++ lib/zendesk2/help_center/get_help_center_section.rb
