+++ lib/zendesk2/help_center/get_help_center_sections_articles.rb
