+++ lib/zendesk2/help_center/search_help_center_articles.rb
