+++ lib/zendesk2/help_center/section.rb
