+++ lib/zendesk2/help_center/sections.rb
