+++ lib/zendesk2/help_center/update_help_center_section.rb
